Move refresh_order_base_prices functions from discount app to order app
